requirements.txt
tensorflow
pillow
scikit-learn
opencv-python
imutils
numpy
matplotlib